# packages/shared/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/taskeroo/tasks": {
      "post": {
        "operationId": "TaskerooController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Create a new task",
        "tags": [
          "Task"
        ]
      },
      "get": {
        "operationId": "TaskerooController_listTasks",
        "parameters": [
          {
            "name": "assignee",
            "required": false,
            "in": "query",
            "description": "Filter tasks by assignee name",
            "schema": {
              "example": "AgentAlpha",
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": false,
            "in": "query",
            "description": "Filter tasks by session ID",
            "schema": {
              "example": "session-123-abc",
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "description": "Filter tasks by tag name",
            "schema": {
              "example": "bug",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (1-100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "List tasks with optional filtering and pagination",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}": {
      "patch": {
        "operationId": "TaskerooController_updateTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Update task description",
        "tags": [
          "Task"
        ]
      },
      "delete": {
        "operationId": "TaskerooController_deleteTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Delete a task",
        "tags": [
          "Task"
        ]
      },
      "get": {
        "operationId": "TaskerooController_getTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Get a task by ID",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}/assign": {
      "patch": {
        "operationId": "TaskerooController_assignTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Assign a task to someone",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}/comments": {
      "post": {
        "operationId": "TaskerooController_addComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Add a comment to a task",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}/status": {
      "patch": {
        "operationId": "TaskerooController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeTaskStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid status transition or comment required"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Change task status",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}/tags": {
      "post": {
        "operationId": "TaskerooController_addTagToTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task UUID",
            "schema": {
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag added to task successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Add a tag to a task",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/{id}/tags/{tagId}": {
      "delete": {
        "operationId": "TaskerooController_removeTagFromTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tag removed from task successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Task not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Remove a tag from a task",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/tags": {
      "post": {
        "operationId": "TaskerooController_createTag",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tag created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Create a new tag",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/tags/all": {
      "get": {
        "operationId": "TaskerooController_getAllTags",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all tags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Get all tags",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/tags/{tagId}": {
      "delete": {
        "operationId": "TaskerooController_deleteTag",
        "parameters": [
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Tag deleted successfully"
          },
          "404": {
            "description": "Tag not found"
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "summary": "Delete a tag from the system",
        "tags": [
          "Task"
        ]
      }
    },
    "/api/v1/taskeroo/tasks/mcp": {
      "get": {
        "operationId": "TaskerooController_handleMcp_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "post": {
        "operationId": "TaskerooController_handleMcp_post",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "put": {
        "operationId": "TaskerooController_handleMcp_put",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "delete": {
        "operationId": "TaskerooController_handleMcp_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "patch": {
        "operationId": "TaskerooController_handleMcp_patch",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "options": {
        "operationId": "TaskerooController_handleMcp_options",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "head": {
        "operationId": "TaskerooController_handleMcp_head",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      },
      "search": {
        "operationId": "TaskerooController_handleMcp_search",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-Cookie": []
          }
        ],
        "tags": [
          "Task"
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Monorepo API",
    "description": "The AI Monorepo API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-Cookie": {
        "type": "http",
        "in": "cookie",
        "name": "access_token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the task",
            "example": "Implement user authentication"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the task",
            "example": "Add JWT-based authentication to the API"
          },
          "assignee": {
            "type": "string",
            "description": "Name of the assignee (for AI agents)",
            "example": "AgentAlpha"
          },
          "sessionId": {
            "type": "string",
            "description": "Session ID for tracking AI agent work",
            "example": "session-123-abc"
          },
          "tagNames": {
            "description": "Array of tag names to associate with the task",
            "example": [
              "bug",
              "urgent"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dependsOnIds": {
            "description": "Array of task IDs that this task depends on",
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
...
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the task",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "description": "Name of the task",
            "example": "Implement user authentication"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the task",
            "example": "Add JWT-based authentication to the API"
          },
          "status": {
            "type": "string",
            "description": "Current status of the task",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "FOR_REVIEW",
              "DONE"
            ],
            "example": "NOT_STARTED"
          },
          "assignee": {
            "type": "string",
            "description": "Name of the assignee (for AI agents)",
            "example": "AgentAlpha",
            "nullable": true
          },
          "sessionId": {
            "type": "string",
            "description": "Session ID for tracking AI agent work",
            "example": "session-123-abc",
            "nullable": true
          },
          "comments": {
            "description": "Comments associated with the task",
            "example": [
              {
                "id": "cmt-001",
                "taskId": "123e4567-e89b-12d3-a456-426614174000",
                "commenterName": "John Doe",
                "content": "Please prioritize this task.",
                "createdAt": "2025-11-03T11:00:00.000Z"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            }
          },
          "tags": {
            "description": "Tags associated with the task",
            "example": [
              {
                "id": "tag-001",
                "name": "bug",
                "color": "#FF5733",
                "description": "Issues that need to be fixed",
                "createdAt": "2025-11-03T10:00:00.000Z",
                "updatedAt": "2025-11-03T10:00:00.000Z"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagResponseDto"
            }
          },
          "createdBy": {
            "type": "string",
            "description": "Name of the person who created the task",
            "example": "Fran"
          },
          "dependsOnIds": {
            "description": "Array of task IDs that this task depends on",
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "description": "Task creation timestamp",
            "example": "2025-11-03T10:30:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Task last update timestamp",
            "example": "2025-11-03T12:45:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "comments",
          "tags",
          "createdBy",
          "dependsOnIds",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the task",
            "example": "Implement user authentication"
          },
          "description": {
            "type": "string",
            "description": "Detailed description of the task",
            "example": "Add JWT-based authentication to the API"
          },
          "assignee": {
            "type": "string",
            "description": "Name of the assignee (for AI agents)",
            "example": "AgentAlpha"
          },
          "sessionId": {
            "type": "string",
            "description": "Session ID for tracking AI agent work",
            "example": "session-123-abc"
          },
          "tagNames": {
            "description": "Array of tag names to associate with the task",
            "example": [
              "bug",
              "urgent"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "dependsOnIds": {
            "description": "Array of task IDs that this task depends on",
            "example": [
              "uuid-1",
              "uuid-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TaskListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "List of tasks",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of tasks matching the filters",
            "example": 42
          },
          "page": {
            "type": "number",
            "description": "Current page number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Number of items per page",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "description": "Total number of pages",
            "example": 3
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
    }
  }
}